pyproject.toml
[tool.poetry]
name = "aws"
version = "0.1.0"
description = ""
authors = ["nugzari.rostiashvili.1 <nugzari.rostiashvili.1@btu.edu.ge>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.42.72"
python-dotenv = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
